requirements.txt
request
requests_oauthlib
azure-identity
msgraph-sdk
msal